+++ .github/workflows/check.yml
@@ -13,4 +13,5 @@ jobs:
               uses: actions/checkout@v3
             - uses: ./.github/actions/pnpm-setup
             - uses: ./.github/actions/lint
+            - uses: ./.github/actions/test
             - uses: ./.github/actions/build-example
