+++ package.json
@@ -13,6 +13,7 @@
     "clean": "turbo run clean && rm -rf node_modules pnpm-lock.yaml",
     "lint": "eslint --ext .js,.ts,.tsx .",
     "lint:fix": "eslint --fix --ext .js,.ts,.tsx .",
+    "test": "turbo run test",
     "e2e:test": "turbo run e2e:test",
     "version": "./.changeset/version",
     "release": "./.changeset/release",
