+++ packages/tests-unit/package.json
@@ -1,21 +1,23 @@
 {
   "name": "tests-unit",
   "private": true,
+  "type": "module",
   "scripts": {
     "clean": "rm -rf .turbo && rm -rf node_modules",
     "dev": "vitest",
     "test": "vitest run --coverage"
   },
   "dependencies": {
-    "@opennextjs/aws": "workspace:*",
-    "@open-next/utils": "workspace:*"
+    "@open-next/utils": "workspace:*",
+    "@opennextjs/aws": "workspace:*"
   },
   "devDependencies": {
     "@types/testing-library__jest-dom": "^5.14.9",
-    "@vitest/coverage-v8": "^0.34.1",
+    "@vitest/coverage-v8": "^2.1.3",
     "jsdom": "^22.1.0",
-    "vitest": "^0.34.1",
-    "vite": "4.4.9"
+    "vite": "5.4.9",
+    "vite-tsconfig-paths": "^5.0.1",
+    "vitest": "^2.1.3"
   },
   "version": null
 }
