--- packages/tests-unit/tests/example.test.ts
@@ -1,7 +1,0 @@
-import { add } from "@open-next/utils";
-
-describe("test", () => {
-  it("passes", () => {
-    expect(add(34.5, 34.5)).toEqual(69);
-  });
-});
