+++ packages/tests-unit/vitest.config.ts
@@ -1,10 +1,9 @@
-/** @type {import('vite').UserConfig} */
-
-import { defineConfig } from "vite";
+import tsconfigPaths from "vite-tsconfig-paths";
+import { defineConfig } from "vitest/config";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [],
+  plugins: [tsconfigPaths()],
   test: {
     globals: true,
     environment: "node",
